backend/app/services/recipes_model/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recipes-rag"
version = "0.1.0"
description = "RAG pipeline for recipe retrieval (CSV -> JSONL -> FAISS -> search)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0",
  "sentence-transformers>=2.7",
  "faiss-cpu>=1.7.4",
  "python-dotenv>=1.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = []

[tool.setuptools]
packages = []
